new full vehicle restrict, ammo rebalancing for supply boxes
making the supply boxes last a little longer now that base ammobox is
not replenning. needs a bit more work. full vehicle now works
differently with some sanity checks alongside, with the aim of
streamlining slightly and making it work on the hind. Also adding
littlebirds to restriction system so that it can't be abused by infantry
if we spawn them to let loose after a long night